OSIS-3949 removed 'contrib' module and moved 'BusinessValidationMessage' and 'BusinessValidator' to base/ddd/utils
