+++ base/ddd/utils/validation_message.py
@@ -23,25 +23,44 @@
 #    see http://www.gnu.org/licenses/.
 #
 ##############################################################################
-from django.conf import settings
-from django.utils.translation import gettext_lazy as _
-
-from program_management.ddd.business_types import *
-from base.ddd.utils.business_validator import BusinessValidator
+from enum import Enum
+from typing import List
 
 
-# Implemented from AttachPermission._check_year_is_editable
-class MinimumEditableYearValidator(BusinessValidator):
-    def __init__(self, tree: 'ProgramTree', node_to_add: 'Node', path: 'Path'):
-        super(MinimumEditableYearValidator, self).__init__()
-        self.tree = tree
-        self.node_to_add = node_to_add
-        self.path = path
+class MessageLevel(Enum):
+    WARNING = "warning"
+    ERROR = "error"
+    INFO = "info"
+    SUCCESS = "succes"
 
-    def validate(self):
-        if self.tree.root_node.year < settings.YEAR_LIMIT_EDG_MODIFICATION:
-            self.add_error_message(
-                _("Cannot perform action on a education group before %(limit_year)s") % {
-                    "limit_year": settings.YEAR_LIMIT_EDG_MODIFICATION
-                }
-            )
+    # Se baser sur le Message level de python
+
+
+class BusinessValidationMessage:
+    message = None
+    level = MessageLevel.ERROR
+
+    def __init__(self, message: str, level: MessageLevel = None):
+        self.message = message
+        self.level = level
+
+    def __eq__(self, other):
+        if isinstance(other, str):
+            return other == self.message
+        return other.message == self.message
+
+    def __str__(self):
+        return "%(level)s %(msg)s" % {'level': self.level, 'msg': self.message}
+
+    def is_error(self):
+        return self.level == MessageLevel.ERROR
+
+    def is_success(self):
+        return self.level == MessageLevel.SUCCESS
+
+    def is_warning(self):
+        return self.level == MessageLevel.WARNING
+
+    @staticmethod
+    def contains_errors(messages: List['BusinessValidationMessage']) -> bool:
+        return any(msg.is_error() for msg in messages)
