+++ program_management/ddd/validators/_attach_finality_end_date.py
@@ -29,7 +29,7 @@ from django.utils.translation import ngettext
 from base.models.enums.education_group_types import TrainingType
 from program_management.ddd.business_types import *
 from program_management.ddd.domain import program_tree
-from program_management.ddd.contrib.validation import BusinessValidator
+from base.ddd.utils.business_validator import BusinessValidator
 
 
 # Implemented from _check_end_year_constraints_on_2m
