+++ program_management/ddd/validators/_authorized_relationship.py
@@ -26,7 +26,7 @@
 from django.utils.translation import gettext as _
 
 from program_management.ddd.business_types import *
-from program_management.ddd.contrib.validation import BusinessValidator
+from base.ddd.utils.business_validator import BusinessValidator
 
 
 # Implemented from CheckAuthorizedRelationship (management.py)
