+++ program_management/ddd/validators/_node_duplication.py
@@ -26,7 +26,7 @@
 from django.utils.translation import gettext_lazy as _
 
 from program_management.ddd.business_types import *
-from program_management.ddd.contrib.validation import BusinessValidator
+from base.ddd.utils.business_validator import BusinessValidator
 
 
 #  Implemented from _check_new_attach_is_not_duplication
