+++ program_management/ddd/validators/_parent_child_academic_year.py
@@ -26,7 +26,7 @@
 from django.utils.translation import gettext_lazy as _
 
 from program_management.ddd.business_types import *
-from program_management.ddd.contrib.validation import BusinessValidator
+from base.ddd.utils.business_validator import BusinessValidator
 
 
 # Implemented from GroupElementYear._check_same_academic_year_parent_child_branch
