+++ program_management/tests/ddd/domain/test_program_tree.py
@@ -25,7 +25,7 @@
 ##############################################################################
 from django.test import SimpleTestCase
 
-from program_management.ddd.contrib.validation import MessageLevel
+from base.ddd.utils.validation_message import MessageLevel
 from program_management.ddd.domain import node
 from program_management.ddd.domain.node import Node, NodeGroupYear
 from program_management.ddd.domain.program_tree import ProgramTree
