+++ program_management/tests/ddd/service/test_attach_node_service.py
@@ -30,7 +30,7 @@ from django.utils.translation import gettext as _
 
 from base.models.enums.education_group_types import TrainingType
 from base.models.enums.link_type import LinkTypes
-from program_management.ddd.contrib.validation import MessageLevel, BusinessValidationMessage
+from base.ddd.utils.validation_message import MessageLevel, BusinessValidationMessage
 from program_management.ddd.domain import program_tree
 from program_management.ddd.service import attach_node_service
 from program_management.ddd.validators._attach_finality_end_date import AttachFinalityEndDateValidator
