+++ program_management/views/tree/attach.py
@@ -40,7 +40,7 @@ from base.views.common import display_warning_messages, display_success_messages
     display_business_messages
 from base.views.mixins import AjaxTemplateMixin
 from program_management.business.group_element_years import management
-from program_management.ddd.contrib.validation import BusinessValidationMessage
+from base.ddd.utils.validation_message import BusinessValidationMessage
 from program_management.ddd.service import attach_node_service
 from program_management.forms.tree.attach import AttachNodeForm, AttachNodeFormSet
 from program_management.models.enums.node_type import NodeType
